+++ tests/macrotest.py
@@ -1607,6 +1607,27 @@ class CommonSkoolMacroTest:
         self._assert_error(writer, '#SPACE({no})', "Unrecognised field 'no': {no}", prefix)
         self._assert_error(writer, '#SPACE({foo)', "Invalid format string: {foo", prefix)
 
+    def test_macro_str(self):
+        writer = self._get_writer(snapshot=[ord(c) for c in 'Hello'])
+
+        self.assertEqual(writer.expand('#STR0,2'), 'He')
+        self.assertEqual(writer.expand('#STR(0,5)'), 'Hello')
+        self.assertEqual(writer.expand('#STR(0+1,4)'), 'ello')
+        self.assertEqual(writer.expand('#LET(a=3)#STR({a},2)'), 'lo')
+        self.assertEqual(writer.expand('#STR(0,#IF(1)(4))'), 'Hell')
+
+    def test_macro_str_invalid(self):
+        writer = self._get_writer()
+        prefix = ERROR_PREFIX.format('STR')
+
+        self._assert_error(writer, '#STR', "No parameters (expected 2)", prefix)
+        self._assert_error(writer, '#STR2', "Not enough parameters (expected 2): '2'", prefix)
+        self._assert_error(writer, '#STR(1,2,3)', "Too many parameters (expected 2): '1,2,3'", prefix)
+        self._assert_error(writer, '#STR(2', "No closing bracket: (2", prefix)
+        self._assert_error(writer, '#STR(0,5$3)', "Cannot parse integer '5$3' in parameter string: '0,5$3'", prefix)
+        self._assert_error(writer, '#STR({no})', "Unrecognised field 'no': {no}", prefix)
+        self._assert_error(writer, '#STR({foo)', "Invalid format string: {foo", prefix)
+
     def test_macro_udg_invalid(self):
         writer = self._get_writer(snapshot=[0] * 8)
         prefix = ERROR_PREFIX.format('UDG')